+++ frontend/appflowy_flutter/lib/plugins/database/widgets/row/row_document.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/plugins/database/grid/application/row/row_document_bloc.dart';
 import 'package:appflowy/plugins/document/application/document_bloc.dart';
+import 'package:appflowy/plugins/document/presentation/editor_drop_handler.dart';
 import 'package:appflowy/plugins/document/presentation/editor_drop_manager.dart';
 import 'package:appflowy/plugins/document/presentation/editor_page.dart';
 import 'package:appflowy/plugins/document/presentation/editor_style.dart';
@@ -45,7 +46,7 @@ class RowDocument extends StatelessWidget {
               ),
             ),
             finish: () => _RowEditor(
-              viewPB: state.viewPB!,
+              view: state.viewPB!,
               onIsEmptyChanged: (isEmpty) => context
                   .read<RowDocumentBloc>()
                   .add(RowDocumentEvent.updateIsEmpty(isEmpty)),
@@ -59,11 +60,11 @@ class RowDocument extends StatelessWidget {
 
 class _RowEditor extends StatelessWidget {
   const _RowEditor({
-    required this.viewPB,
+    required this.view,
     this.onIsEmptyChanged,
   });
 
-  final ViewPB viewPB;
+  final ViewPB view;
   final void Function(bool)? onIsEmptyChanged;
 
   @override
@@ -81,7 +82,7 @@ class _RowEditor extends StatelessWidget {
       //
       create: (_) => EditorDropManagerState(),
       child: BlocProvider(
-        create: (context) => DocumentBloc(documentId: viewPB.id)
+        create: (context) => DocumentBloc(documentId: view.id)
           ..add(const DocumentEvent.initial()),
         child: BlocConsumer<DocumentBloc, DocumentState>(
           listenWhen: (previous, current) =>
@@ -114,22 +115,27 @@ class _RowEditor extends StatelessWidget {
 
             return Consumer<EditorDropManagerState>(
               builder: (_, dropState, __) => BlocProvider<ViewInfoBloc>(
-                create: (context) => ViewInfoBloc(view: viewPB),
+                create: (context) => ViewInfoBloc(view: view),
                 child: IntrinsicHeight(
                   child: Container(
                     constraints: const BoxConstraints(minHeight: 300),
-                    child: AppFlowyEditorPage(
-                      shrinkWrap: true,
-                      autoFocus: false,
+                    child: EditorDropHandler(
+                      viewId: view.id,
                       editorState: editorState,
-                      styleCustomizer: EditorStyleCustomizer(
-                        context: context,
-                        padding: const EdgeInsets.only(left: 16, right: 54),
+                      isDropEnabled: dropState.isDropEnabled,
+                      child: AppFlowyEditorPage(
+                        shrinkWrap: true,
+                        autoFocus: false,
+                        editorState: editorState,
+                        styleCustomizer: EditorStyleCustomizer(
+                          context: context,
+                          padding: const EdgeInsets.only(left: 16, right: 54),
+                        ),
+                        showParagraphPlaceholder: (editorState, node) =>
+                            editorState.document.isEmpty,
+                        placeholderText: (node) =>
+                            LocaleKeys.cardDetails_notesPlaceholder.tr(),
                       ),
-                      showParagraphPlaceholder: (editorState, node) =>
-                          editorState.document.isEmpty,
-                      placeholderText: (node) =>
-                          LocaleKeys.cardDetails_notesPlaceholder.tr(),
                     ),
                   ),
                 ),
